internal/database/own_signal_configurations.go
pbckbge dbtbbbse

import (
	"context"

	"github.com/keegbncsmith/sqlf"
	"github.com/lib/pq"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type SignblConfigurbtion struct {
	ID                   int
	Nbme                 string
	Description          string
	ExcludedRepoPbtterns []string
	Enbbled              bool
}

type SignblConfigurbtionStore interfbce {
	LobdConfigurbtions(ctx context.Context, brgs LobdSignblConfigurbtionArgs) ([]SignblConfigurbtion, error)
	IsEnbbled(ctx context.Context, nbme string) (bool, error)
	UpdbteConfigurbtion(ctx context.Context, brgs UpdbteSignblConfigurbtionArgs) error
	WithTrbnsbct(context.Context, func(store SignblConfigurbtionStore) error) error
}

type UpdbteSignblConfigurbtionArgs struct {
	Nbme                 string
	ExcludedRepoPbtterns []string
	Enbbled              bool
}

type signblConfigurbtionStore struct {
	*bbsestore.Store
}

func SignblConfigurbtionStoreWith(store bbsestore.ShbrebbleStore) SignblConfigurbtionStore {
	return &signblConfigurbtionStore{Store: bbsestore.NewWithHbndle(store.Hbndle())}
}

func (s *signblConfigurbtionStore) With(other bbsestore.ShbrebbleStore) *signblConfigurbtionStore {
	return &signblConfigurbtionStore{s.Store.With(other)}
}

type LobdSignblConfigurbtionArgs struct {
	Nbme string
}

func (s *signblConfigurbtionStore) LobdConfigurbtions(ctx context.Context, brgs LobdSignblConfigurbtionArgs) ([]SignblConfigurbtion, error) {
	q := "SELECT id, nbme, description, excluded_repo_pbtterns, enbbled FROM own_signbl_configurbtions %s ORDER BY id;"

	where := sqlf.Sprintf("")
	if len(brgs.Nbme) > 0 {
		where = sqlf.Sprintf("WHERE nbme = %s", brgs.Nbme)
	}

	multiScbn := bbsestore.NewSliceScbnner(func(scbnner dbutil.Scbnner) (SignblConfigurbtion, error) {
		vbr temp SignblConfigurbtion
		err := scbnner.Scbn(
			&temp.ID,
			&temp.Nbme,
			&temp.Description,
			pq.Arrby(&temp.ExcludedRepoPbtterns),
			&temp.Enbbled,
		)
		if err != nil {
			return SignblConfigurbtion{}, err
		}
		return temp, nil
	})

	return multiScbn(s.Query(ctx, sqlf.Sprintf(q, where)))
}

func (s *signblConfigurbtionStore) IsEnbbled(ctx context.Context, nbme string) (bool, error) {
	configurbtions, err := s.LobdConfigurbtions(ctx, LobdSignblConfigurbtionArgs{Nbme: nbme})
	if err != nil {
		return fblse, err
	} else if len(configurbtions) == 0 {
		return fblse, errors.New("signbl configurbtion not found")
	}
	return configurbtions[0].Enbbled, nil
}

func (s *signblConfigurbtionStore) UpdbteConfigurbtion(ctx context.Context, brgs UpdbteSignblConfigurbtionArgs) error {
	q := "UPDATE own_signbl_configurbtions SET enbbled = %s, excluded_repo_pbtterns = %s WHERE nbme = %s"
	return s.Exec(ctx, sqlf.Sprintf(q, brgs.Enbbled, pq.Arrby(brgs.ExcludedRepoPbtterns), brgs.Nbme))
}

func (s *signblConfigurbtionStore) WithTrbnsbct(ctx context.Context, f func(store SignblConfigurbtionStore) error) error {
	return s.Store.WithTrbnsbct(ctx, func(tx *bbsestore.Store) error {
		return f(s.With(tx))
	})
}
